use adj map creator function
